backend/app/storage.py
import boto3
import os
from dotenv import load_dotenv  
load_dotenv()

s3 = boto3.client(
    "s3",
    endpoint_url=os.getenv("MINIO_ENDPOINT"),
    aws_access_key_id=os.getenv("MINIO_ROOT_USER"),
    aws_secret_access_key=os.getenv("MINIO_ROOT_PASSWORD"),
    region_name="us-east-1"
)

BUCKET_NAME = os.getenv("MINIO_BUCKET")
