backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
celery[redis]
python-dotenv
pydantic
minio
requests
boto3
botocore
# manim
Pillow
